fix silent stack overflow

+++ src/boot/ksdk1.1.0/fft.h
@@ -5,6 +5,4 @@
 #include <stdint.h>
 #include <complex.h>
 
-//void fft_slow(int* x, double complex* X, uint16_t N);
-
-void fft(int* x, double complex* X, uint16_t N);
+void fft(int* x, float complex* X, uint16_t N);
